Boolean logic bug fixes and string slicing operations
Bug fixes related to handling of 'or' and 'and' operators

Fix incorrect handling of 'and' and 'or' in python, and bad conversion to bool logic.

Signed-off-by: Amichai <amichai@ufora.com>

more string slicing logic and testing

Signed-off-by: Amichai <amichai@ufora.com>

Made the string slicing tests use the async executor API

This allows our remote unit tests to finish in a reasonable amount of time.