Ensure that generated classmembers return the correct kind of None at the end.
By default, generated FORA code will produce a FORA 'nothing' if we don't
have a return statement. This is not the same as a PyNone object, so we need
to ensure that all function bodies have a 'return PyNone(nothing)' at the
end. We were missing such a statement for generated class members.