Basic model for generator expressions.
If a user writes `list(expr for x in iterable)`, we convert the generator
expression into a "PyforaGenerator" object. This behavior is controllable
by the 'iterable', with a default implementation that just falls back
to the standard iterator.

The point of allowing 'iterable' to control this behavior is that we
want collections that have strong structure (xrange, lists, etc) to be
able to expose that structure to algorithms that can consume generators
in a parallel way.