fix bug in PyObjectBase:
also fixup some formatting. The bug was in (fora) lines like
if (a.__le__(b)) ... . Things like a.__le__(b) return
PyBools, but in Fora, the only values which are considered
true in an if expression are nonzero integers -- all classes
and such evaluate to false EVEN IF they implement an
convert (Bool)  method.