Generic model for PyGeneratorObject.
In Pyfora we provide a more general notion of an 'iterator' which is
a "pyfora generator". A pyfora generator is always an iterator, but
it exposes methods to allow users to map, filter, and (optionally) split
it. These can be consumed by operators like sum, list, etc, which
in pyfora can have implementations that take advantage of this infrastructure,
but which fall back to regular iterators.

Generator expressions (and eventually list comprehensions) are implemented
in terms of this pattern. E.g. if a user writes

	sum(isPrime(x) for x in xrange(1000000))

then we actually generate code as follows:

	sum(xrange(1000000).__pyfora_generator__().map(lambda x: isPrime(x)))

this defers control to 'xrange' to decide how to implement 'map', 'filter', and
'split' operations. The sum function converts its arguments to a
pyfora generator and uses that to implement summation.