Getting rid of virtualenv in tests
pyfora is now installed as part of global test setup.
We used to install pyfora in a virtual env because we had a version
conflict with the 'requests' module that prevented tests that used REST
API from passing with the latest version.
The REST API is now gone and so are those tests.