VectorHandle::visitAnyLoadedValues returns whether it visits _all_ data.
We rely on this function to tell us whether we've visited data in the vector.
It's possible for the front of the vector (pagelet and unpaged values) to be populated
but the paged data to not be loaded. In this case, the function was returning
'true', causing the caller to assume it had seen all the data when in fact
it had only seen the tail.

This was, in turn, causing functions in ComputedValueGateway to fail to
load data - they would think data was loaded, set a flag to 'True', and then
later ignore when the flag was reset by the data-loading thread. As a result,
they would fail to trigger the SWO subscription and the system would just hang.