Wrap all internally produced objects in a PyBoundMethod class.
In general, we need to ensure that all objects that user python
code can interact with have PyObjectBase as a mixin. This ensures
that we throw appropriate exception types (e.g. AttributeError,
instead of a native FORA exception). This also gives us a pathway
for later implementing correct stringification on these method
objects.