[#47] An integer raised to an integer power returns an integer
An integer raised to an integer power returns an integer
in python, but a float in fora. I added a cast to Int64 in the
PyInt.fora class which solves this problem. I also modified the
BinaryOperators unit test to check the return types of each operation
in addition to checking equality. Now we are asserting the correct
return type for all of our binary operator tests.