Improve the internal error pathway when generated FORA code has free variables.
This should never come up in a correct usage of the pyfora package, but
during development we can occasionally get into a state where the generated
FORA code has free variables that pyfora doesn't think are free.

This commit detects this case and returns an error message to the front-end
as a PythonToForaConversionError (which is actually comprehensible)
instead of throwing an exception that propagates as a SubscribableWebObject exception.