Fix #89. Ensure that we can keyboard-break in pyfora with blocks.
The futures function 'result' is blocking, so if a user hits
ctrl-c while we're calculating, the exception doesn't propagate
until we're done running the code. This change causes the future
to poll 'result' periodically, and also adds a pathway for
the WBE to throw the exception on exit from the with block.

In the future, we should consider whether we want this to affect
prioritization as well.