Fix #68 Adding checks for disallowed 'return' and 'yield' expressions in with-blocks.
Adding a new visitor that counts yield expressions and return statements and provides
some functionality to query this information (and also did some minor code shifting
between NodeVisitorBases and PyAstUtil to avoid having to create a PyAstVisitorUtil
file for the new analysis to avoid a cyclic import dependency).

Also, added checks in the PyObjectWalker function that deals with pyfora with-blocks
to implement the aforementioned checks. And added tests.

Finally, tidied up the exception handling code in WithBlockExecutor to properly
propagate all exceptions, not only the ones we've thought about so far. This is
done by re-raising all exceptions and by swallowing (returning True in __exit__)
only when the WithBlockCompleted exception is thrown.