Fix #67 - tuple assignment works.
In the case of a proper "tuple assignment" pattern, such as `x,y,z = rhs`,
the AstConverter produces a pattern-match like

	rhs`(`Match, (#TupleAssign, 3))

where 3 is the desired arity of the tuple match (3 here because `x,y,z` has
three elements).

This is handled in two ways - PyTuple can expand its tuple elements directly,
which is very efficient within the compiler. All other objects inherit
from the PyObjectBase mixin a pattern-match that attempts to iterate over the
object to produce the required number of elements, and throws an exception
if it doesn't get the right number.