Fix #93, we can execute code in the repl.
Note - to make this work, we have to be running in a TTY. So the test
harness uses pexpect and is a little awkward.

Essentially, we use the readline module to be able to track all the
commands a user has entered over time. When we see a stackframe or a
function that says its from "<stdin>", we assume its content is somewhere
in 'readline'. Line numbers start at 1 again for each readline
block, so we have to figure out which function we're actually looking
at.

The 'StdinCache' object parses each block of readline objects as
they come in, and grabs all the code objects produced. When we have
a 'with' block get executed (or when we access a function object),
we check the code object against all the ones we've parsed. This
allows us to find the relevant source code without having to actually
have valid line numbers.

Note that we're not doing a perfect job simulating how the interpreter
parses lines - some combination of trailing backslashes, whitespace,
and invalid characters could break the grouping mechanism in
the StdinCache object.