Fix #111, #29, #83. Trigger exceptions when we access unbound variables.
Instead of trying to decide whether a value is "unassigned" at parse time,
we now take all values that get assigned in a scope and initialize them
to a symbol `PyforaUnassignedVariable. We check whether the value is this
variable on reads, and throw an exception if so, so that we can accurately
mimic python's behavior. Note that this obviates the need for calculating
the possibly assigned variables - it's enough to know in a scope whether
a variable is read or written from in the local scope.

In the future, we may want to do some clever analysis to figure out
whether a value is always assigned. This will get rid of the need to
check each time. The compiler can usually figure this out, but it
can place quite a load on it to do so, when we can probably do it from
a simpler pass over the code.