Implemented flatten and tolist on the numpy array
Flatten converts the array to 1-d array and tolist
converts an n-dimensional numpy array to an n-dimensional
list of lists.

Also, you can now np.dot a 1-d array with a 2-d array we'll
reshape the 1-d array to 2-d. This is what numpy does

Implemented the iterator pattern on the numpy array wrapper class