refacting PyObjectWalker, and #120
* Substantially simplified this class. Got rid of pyObjectNodes, and
ObjectVisitors.

* fixes #120 along the way

* Note that Executor.py *still* creates a new PyObjectWalker on each call to
define. Holding a single executor around breaks tests! The reason
is that PyObjectWalker caches the python objects it visits based
on python id, which is apparently a memory location, and this can
get reused. To see this behavior, add in a single PyObjectWalker to
Executor and run `python test.py -py -filter=PurePython -filter=test_int`.
(The three functions `f` in there can sometimes get the same memory address.)
If we want to do better caching in Executor, we'll may have to find a
suitable substitute for python `id`.

* also some general refactoring: renaming in `TypeDescription`,
chop long lines in ExecutorTestCases, ... possibly more