Fix #30 - don't allow 'return' statements in '__init__'
Initializing classes in python is expressed as a mutable operation. To allow
the normal python idiom, we modify user __init__ code to turn assignments to
'self.X' (for various X) into regular assignments to variables self_X, which
we then collect at the end of the function as the members of the new instance.

Rhis transformation is very sensitive, since we're not really modifying 'self'.
In the future, we might be able to do something more sophisticated, to allow
a wider range of behaviors in self. But for now, we're simply turning off
anything other than very simple assignments, which is what we do if the user
tries to 'return' in an __init__. Before this change you'd call the constructor
and not even get an instance back.