Fix #90 User-friendly errors with traces when failing to resolve free-variables.
Collecting the necessary information to produce informative errors when we cannot
resolve a free variable. To achieve that we modify the free-variables and bound-
variables analyses to return sets of (var, location) pairs, where location is a
(line, column) pair indicating where each free-variable, free-variable-chain, or bound-
variable appeared in a file. Furthermore, we throw an UnresolvedFreeVariableException
and catch it higher where we have the remaining info (path and function context) for
the user-facing exception. At this point we throw a UnresolvedFreeVariableExceptionWithTrace
and we add to this trace for each free-var on the path to the free-var which we could
not resolve. That way we build an informative trace for the user. Finally, in
Executor, which is the single entry-point of PyObjectWalker.walkPyObject, we raise
a PythonToForaConversionError if we catch an UnresolvedFreeVariableExceptionWithTrace.