Fix #59 - empty generator functions.
In the core FORA model for yield, which we had just borrowed, if you
yield, you return a tuple of (value, continuation). If you run off the
end of the function, you return nothing. The iterator contract in FORA
specifies that this is OK, so that 'nothing' is the empty iterator.
That's not true in python, so we'd run off the end of a generator function,
return None, and get something we couldn't iterate over.

This change actually checks whether functions have 'yield' in them, and
if so returns a Generator object with the correct contract for python.