fix a bug in rebindFreeVariableMemberAccessChains:
before this change, we weren't searching all member access prefixes
in a chained member expression: only the deepest in
Fora::rebindFreeMemberAccessChains. So, for example, in python,
if we had an expression `e` representing 'x.y.z.w.q' and we called
`e.rebindFreeVariableMemberAccessChain(['x', 'y', 'z'], 'x_y_z')`,
we wouldn't have caught the 'x.y.z' prefix, since we we only looking
at the full chain 'x.y.z.w.q'

This change corrects that, but note that it's currently a (worst case)
n^2 search in the length of a chain. For example, in
Member(Member(Member(x,y),z),w), at each submember we compute
the longest subchain it holds by a linear scan
(in RebindFreeMemberAccesses.cppml:extractChain). We could do better,
but these things should be very short and this is good enough for now.