Drop values from SimpleOfflineCache in LRU order.
The SimpleOfflineCache is a disk cache for cumulus that's really only used in testing.
In some of our tests, we want to be able to run calculations on loop that produce
lots of temporary data, and we don't want the data to get accumulated in memory, so
it needs to get dumped to the offline cache. The old one would drop values randomly,
which would eventually cause the test to fail because it would dump a value it needed.
With this change, it only dumps values from prior iterations, so it's not a problem.

Over the long run, this will go away, because we're going to get rid of the local caches,
and use the persistent cache instead.