ClusterSimulation calls 'forever' from an OutOfProcessDownloader.
Once the host process is using a lot of ram, it becomes impossible to fork,
and we can't run 'forever'. This solves this problem by creating an
OutOfProcessDownloader at startup and using that to execute forever. The subprocess
doesn't have much memory allocated, so it can fork all it wants.