(#136) rename Pandas.py, Numpy.Py -> PurePandas.py, PureNumpy.py
We were running into problems on Windows (and sometimes OSX), due to
case-insensitive filesystems. In Numpy.py, for example, we were
importing numpy itself, which on those systems got routed to
Numpy.py instead of the system numpy. Renaming these files
relieves this problem

Note though that we have no enforcement of this behavior.