Ensure that parallelism actually works by fixing a bug in AstConversion.
We were using "match" to check variables against symbols in an effort to see
whether they are unbound. This was breaking the CFGSplitter. We will replace
this eventually with the CFGWithFutures, which can handle this. But for now,
we switch this from matching to an apply axiom, which the CFGSplitter understands.