Create "JudgmentBundle" and "SimpleForwardReasoner".
This is a fist pass at infrastructure that can more accurately represent
the dependencies between input and output judgments on functions. The
key idea is to perform symbolic execution on FORA code where the variables
are functions from boolean variables to judgments. This allows us to
characterize algebraically the dependency between the inputs and outputs
of a function. For instance, we can have the function

	fun(x,y) { x + y }

and call it with "JudgmentBundles"

	(if !x0 * else if !x1 {Int64} else 1,
		if !x2 * else if !x3 {Int64} else 2)

and get a resulting JudgmentBundle of

	if !(x0 and x2) * else if !(x1 and x3) {Int64} else 3

this accurately encodes the full spectrum of information we have about
the piece of code based on the input judgments without requiring us
to compute all possible combinations of input values.