Fix #132 Removing some dead code, and moved AST code into pyAst subdirectory.
Also, checked that the following code path cannot be substantially optimized.
The call to PyforaInspect.getsourcelines(pyObject) in PyObjectWalker has currently
little overlap with what PyAstUtil.getSourceFilenameAndText(pyObject) does.
The main opportunity for better performance would be to have a "light" version of
'getsourcelines' that only returns the line number, without the actual lines. But
internally the algorithm needs too keep track of the lines, in order to compute the
line-number, so the savings would be minimal.