[#176] fix bug in `isinstance`
this example

    class C: pass
    isinstance(C(), list)

was raising an exception. The reason was in IsSubclass.fora, which
asks for a type's `__bases__` attribute. Problem was that this
wasn't implemented for user defined classes.

Now, in PythonAstConverter, we fill out this field with an
empty tuple. We'll have to be careful when we start implementing
inheritance to actually fill this guy out.