tweak rebindFreeAs:
if we had an expression: x = expr1 and we want to replace
x's with expr2, we used to paste expr2 in directly for expr1.
Now, if expr2 doesn't have any location metadata, and x does,
we insert that into the pasted-in form of expr2.

This was done to improve certain tracebacks in pyfora arising
from uninitialized variables. The point is, those things are
handled by pasting in (using rebindFreeAs) certain try/catch/throw
expressions which check against `PyforaUninitialized. If the
read variable was uninitialized, we throw an exception, but due
to the old rebindFreeAs, we don't get a stacktrace entry for that
exception, since it had no metadata. Now we get it at the read site.