[#199] PyObjectWalker searches base classes
so now, even though we don't support at the translation layer, if
we have classes like:

    class A: pass

    class B(A): pass

the PyObjectWalker can resolve the free variable A in B's AST.

Note we could do better than we're doing here: we could search the
class variables in the base class. This would be an easy extension.