Compiler Cache.
We cache the ClassMediator->CFG relationship as well as the target objects.
We memoize CFGs, Types, JOVs, and certain kinds of Expressions. This
memoization is necessary to keep disk usage low, but also because
without it we can experience exponential overheads in serialization
and deserialization, as well as in disk usage. As an example,
for test_RegressionTreeFitting_1 the disk usage goes from ~10G when
only memoizing CFGs and Expressions, to ~16MB when also memoizing
Types and JOVs.