fix #203: socketIO exception
This is an unfortunate hack to solve this problem we were hitting.
The problem was we had a race condition in websockets, but we
weren't using their multithreaded mode, and socketIO_client
doesn't allow you a way to set that flag. Here we just reach in
and set the flag. Sigh ...