Implement HomogenousDoubleVectorStash.
DoubleVectorStash is intended to allow us to quickly index into 2d-arrays (e.g. lists
of lists of primitives) without as many boundschecks. Basically, we'll detect lookups
like
	v[ix][ix2]
and replace them with a "DoubleVectorStash", which attemps to compute a rectangular region
within 'v' with boundschecks for both ix and ix2 that are valid. Assuming we stay within
these bounds, we can just directly index into the given datastructure, which is very
fast.

Once this works, it will be a natural stepping-stone towards allowing for auto-transposition
of array data. This will provide a very large speedup for programs that are iterating over
column-major data in a row-major way, which we often want to do in ML applications without
having to explicitly transpose the data in a separate data-structure.