Ensure that codegen for external blocks jumps to an internal copy of the block.
In codegen, we have the notion of "internal" and "external" blocks. External blocks
cannot be modified as heavily during optimization because callers expect to
be able to reach them. As a result, if a loop contains an external block, it's hard
to modify the loop.

This change causes the Reasoner to produce an internal block for every external
block and to insist that we link to the internal block from every loop. This ensures
that downstream passes don't get confused.