Create a notion of "Expansion" call target for native code.
This will allow us to cordon off lib-call code that we don't want
to include in standard optimization passes. We'll expand the given
code at the last possible moment (that is, when flattening the code),
so that we can more easily see the higher-level structure of the code
during earlier passes.