Don't drop values from stackframes that are not used later in the function.
We need to keep these around so that we can get accurate predictions of how
long a given function will compute for. Originally, we implemented this so that
we could release intermediate datasets as we were sorting (using a mergesort).
Now that we have an out-of-band sorting mechanism, this is no longer necessary.