Use Pooled refecounts in Tuple slice axiom.
Our old refcounting system is still in place for some axioms. It would be better
if we used Pooled for basically everything, but we don't. Tom's recent change
to the way we handle argument passing caused us to start doing a lot of tuple
slicing, and this was creating extra refcounts that we don't need. E.g.

  (v,)[0,]

was increffing 'v', even if 'v' was Pooled.

This change fixes that.