fix a bug in the implementation of pyfora sort:
pyfora sort was using the old builtin sorting.sort with the default
less function, fun(x, y) { x < y }. Now for PyObjects, x < y is a
PyBool, not a fora bool. In some places in the sorting code, like
quickSort, we are careful to call Bool on less calls. We really
shouldn't need to do that: it should be the responsibility of the
caller to pass in a less which properly gives bools. That's what
this change does.