Ensure that sorting works when many values are the same. Fixes #275.
We tackle this problem by inserting an extra tuple element into the
value being sorted that contains the index of the value in the original
page. This makes each value unique, at the cost of some additional data
overhead. This isn't the ideal way to solve this problem (really, we would
prefer to actually count how many unique values we get at each split point
and then reproduce those values in the given pages), but this is an
expedient solution that uses the existing infrastructure.