Create "OnDemandMemoryRegions" allowing a clean interface around segmentation faults.
This class allows us to map large regions of memory and map them to smaller regions of
memory that are held in shared-memory segments. The same shared-memory segments can be
mapped to multiple regions in the process address space. This class abstracts away
the details of handling signals, mapping memory, awakening blocked threads, etc.