Changes internal representation for FreeVars analysis
We replace the set of free variables with positions with
a dictionary from free vars to sets of positions.
This allows the isFree(var) call to execute in constant time rather
than in linear time in the size of the free-vars-with-pos set.

Because now we can simply return the keys of the free-vars dicts (which
are unique), we now return a list rather than a set. In the unit-tests
we have to wrap this resulting list into a set to compare with the
expected result without depending on list ordering.