fix a bug in OOPP variable assignments
We had been assuming that any assignment that could conceivably run
in the body of an OOPP withBlock actually happened. This resulted
in cryptic exceptions if any such assignment never took place
(for example, `if False: y = 2` won't run, and if y isn't assigned
somewhere else, y is an UnboundLocal).