Fixes bug in support for vec-of-vec in GPUS and adds a test for it.
We were hoarking the CFG in the ConvertForGPUExecutions conversions.
The first problem was that we were modifying the type of vectors
in HomogenousVectorLookup and made a self-referential type that would
cause trying to print it to explode by running out of stack.

The second problem was that we were replacing the HomogenousVectorLookup
expression with an expression that could potentially be of a different type.