fix #294: NameErrors work properly in OOPP again.
During python object walking, when we hit what would be NameErrors
in execution, we insert into our binary stream an encoding of this fact.
When we deserialize that stream into a fora object, we use that marker
to create appropriate throw expressions. When we deserialize that stream
into a python object, we create a special value for these should-be
NameErrors. When rehdrating functions
(in PythonObjectRehydratorHelpers.instantiateFunction), we filter out
these values in the memberDict, which is what is used for looking up
free variables in the function. This results in NameErrors being raised
as desired.